LVGL moved Freetype embedded in LVGL8
