--- lib/libesp32_lvgl/lv_lib_freetype/library.json
@@ -1,15 +1,0 @@
-{
-  "name":"lv_lib_freetype",
-  "description":"Interface to FreeType to generate font bitmaps run time",
-  "keywords":"lvgl, freetype",
-  "license": "MIT License",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/lvgl/lv_lib_freetype"
-	},
-  "frameworks": "arduino",
-  "platforms": "espressif32",
-  "build": {
-    "flags": [ "-DLV_LVGL_H_INCLUDE_SIMPLE", "-I$PROJECT_DIR/include", "-includetasmota_options.h" ]
-  }
-}
