+++ lib/libesp32_lvgl/lvgl/src/extra/libs/lv_libs.h
@@ -20,7 +20,7 @@ extern "C" {
 // #include "gif/lv_gif.h"
 // #include "qrcode/lv_qrcode.h"
 // #include "sjpg/lv_sjpg.h"
-// #include "freetype/lv_freetype.h"
+#include "freetype/lv_freetype.h"
 // #include "rlottie/lv_rlottie.h"
 
 /*********************
