+++ tasmota/xdrv_52_3_berry_tasmota.ino
@@ -214,6 +214,9 @@ extern "C" {
       be_map_insert_int(vm, "program_free", ESP.getFreeSketchSpace() / 1024);
       be_map_insert_int(vm, "heap_free", ESP_getFreeHeap() / 1024);
       be_map_insert_int(vm, "frag", ESP_getHeapFragmentation());
+      // give info about stack size
+      be_map_insert_int(vm, "stack_size", SET_ESP32_STACK_SIZE / 1024);
+      be_map_insert_int(vm, "stack_low", uxTaskGetStackHighWaterMark(nullptr) / 1024);
       if (UsePSRAM()) {
         be_map_insert_int(vm, "psram", ESP.getPsramSize() / 1024);
         be_map_insert_int(vm, "psram_free", ESP.getFreePsram() / 1024);
